+++ 1.21.6/src/main/java/com/bleudev/ppl_utils/client/ClientCallbacks.java
@@ -1,38 +1,34 @@
 package com.bleudev.ppl_utils.client;
 
 import com.bleudev.ppl_utils.client.compat.modmenu.PplUtilsConfig;
-import com.bleudev.ppl_utils.util.ServerUtils;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.hud.ChatHudLine;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import static com.bleudev.ppl_utils.util.ServerUtils.executeCommand;
+import static com.bleudev.ppl_utils.util.ServerUtils.isLobbyCommandWorking;
+
 public class ClientCallbacks {
-    public static void executeLobby() {
-        var client = MinecraftClient.getInstance();
-        if (ServerUtils.isClientOnServerSupportsLobbyCommand(client))
-            ServerUtils.executeCommand(client, "lobby");
+    public static void executeLobby(@NotNull MinecraftClient client) {
+        if (isLobbyCommandWorking(client))
+            executeCommand(client, "lobby");
     }
 
     @Nullable
-    private static String extractPlayer(String content) {
-        content = content.toLowerCase();
-        if (content.contains(" ")) return extractPlayer(content.replace(" ", ""));
+    private static String extractPlayer(@NotNull String content) {
+        String e = content;
 
-        final int i, j;
-        if (content.startsWith("<") && (i = content.indexOf(">")) != -1 && content.length() >= i + 1)
-            return extractPlayer(content.substring(i+1));
-
-        if (content.startsWith("[+]") || content.startsWith("[-]")) {
-            var user = content.substring(3);
-
-            if (user.startsWith("[")) {
-                if ((j = user.indexOf("]")) != -1 && content.length() >= j + 1) {
-                    var ans = user.substring(j+1);
-                    if (user.startsWith("[" + ans + "head]")) return ans;
-                }
-            } else return user;
+        final var NICKNAME = "[^ \\[\\]]+";
+        e = e.replaceAll(String.format("(<%1$2s>)* *\\[\\+] *\\[%1$2s head] *", NICKNAME), "");
+        e = e.replaceAll(String.format("(<%1$2s>)* *\\[-] *\\[%1$2s head] *", NICKNAME), "");
+        if (content.equals(e)) {
+            e = e.replaceAll(String.format("(<%1$2s>)* *\\[\\+] *", NICKNAME), "");
+            e = e.replaceAll(String.format("(<%1$2s>)* *\\[-] *", NICKNAME), "");
         }
-
+        System.out.println(e);
+        if (e.matches(NICKNAME))
+            return e;
         return null;
     }
 
@@ -46,5 +42,9 @@ public class ClientCallbacks {
                     .map(String::toLowerCase).toList()
                     .contains(user);
         return true;
+    }
+
+    public static boolean shouldRenderLobbyButton(@NotNull MinecraftClient client) {
+        return PplUtilsConfig.lobby_button_enabled && isLobbyCommandWorking(client);
     }
 }